requirements.txt
annotated-types==0.7.0
anyio==4.9.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
psycopg2-binary==2.9.9
python-multipart==0.0.20
reportlab==4.4.2
sniffio==1.3.1
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
pydantic-settings==2.10.1
python-dotenv==1.1.1
alembic==1.16.4
qrcode==8.0
pandas==2.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-ldap==3.4.4

